Mock dates in tests to prevent unexpected behaviour
As the tests rely on functions being called on specific dates, I
have mocked the date function to ensure that the date when the
functions are called is consistent and reliable.

This prevents (an extremely small) edge case where a test could
be called just before a the current date changes, and by the time
the code under test is called, it is a different date.