Refactor date handling
There was an enormous bug that was left over from extracting the
statementPrinter from account where dates were being set in the
printer rather than account.

This was corrected as part of a refactor of date handling in
general. Dates are now stored as part of the account history, and
the statement printer has responsibility for formatting these
appropriately. As part of this I removed the separate todayAsString
module, as it was an unnecessary module. As a single, short
function it sits comfortably within StatementPrinter.