Change responsibility for statement order
The responsibility for ordering the statement in reverse
chronological order was on the account. It makes more sense for
this to sit with the statement printer, as you can imagine
scenarios where different printers might order the history in
different ways.

Because of this, the account now orders its history normally, and
the printer reverses this to print the statement.