Refactored Account tests
Duplicate tests that now belong in Deposit or Withdrawl were
deleted. These were replaced by tests that check whether the
deposit and withdraw methods create Deposit and Withdraw objects
respectively.

This involved injecting transaction types into the Account
function, which I am not overly happy with, as this means that
every time you add a new transaction type, it needs to be added
here to, but it was the only way I could properly unit test the
Account function.