Tidy up code
Any remaining traces of the raw array being used to represent the
transaction list in the Account class were removed.

A couple of tweaks were made to the transaction test descriptions
to make them clearer.

newTransactionList was renamed to transactionList and
newTransactions to transactions in the Statement class. This was
a holdover from when I was refactoring to incorporate the new
TransactionList class.